requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
gradio>=4.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pandas>=2.0.0
pytest>=8.0.0
pytest-playwright>=0.4.0
httpx>=0.27.0
